+++ pinot-broker/src/main/java/org/apache/pinot/broker/api/services/PinotBrokerAppConfigsService.java
@@ -16,32 +16,35 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.pinot.broker.api.resources;
+package org.apache.pinot.broker.api.services;
 
+import io.swagger.annotations.Api;
+import io.swagger.annotations.ApiKeyAuthDefinition;
+import io.swagger.annotations.Authorization;
+import io.swagger.annotations.SecurityDefinition;
+import io.swagger.annotations.SwaggerDefinition;
+import javax.ws.rs.GET;
 import javax.ws.rs.Path;
-import javax.ws.rs.core.Application;
-import javax.ws.rs.core.Context;
-import org.apache.pinot.broker.api.services.PinotBrokerAppConfigsService;
-import org.apache.pinot.broker.broker.BrokerAdminApiApplication;
+import javax.ws.rs.Produces;
+import javax.ws.rs.core.HttpHeaders;
+import javax.ws.rs.core.MediaType;
 import org.apache.pinot.common.utils.PinotAppConfigs;
-import org.apache.pinot.spi.env.PinotConfiguration;
+
+import static org.apache.pinot.spi.utils.CommonConstants.SWAGGER_AUTHORIZATION_KEY;
 
 
 /**
  * Resource to get the app configs {@link PinotAppConfigs} for
  * the broker.
  */
+@Api(tags = "AppConfig", authorizations = {@Authorization(value = SWAGGER_AUTHORIZATION_KEY)})
+@SwaggerDefinition(securityDefinition = @SecurityDefinition(apiKeyAuthDefinitions = @ApiKeyAuthDefinition(name =
+    HttpHeaders.AUTHORIZATION, in = ApiKeyAuthDefinition.ApiKeyLocation.HEADER, key = SWAGGER_AUTHORIZATION_KEY)))
 @Path("/")
-public class PinotBrokerAppConfigs implements PinotBrokerAppConfigsService {
+public interface PinotBrokerAppConfigsService {
 
-  @Context
-  private Application _application;
-
-  @Override
-  public String getAppConfigs() {
-    PinotConfiguration pinotConfiguration =
-        (PinotConfiguration) _application.getProperties().get(BrokerAdminApiApplication.PINOT_CONFIGURATION);
-    PinotAppConfigs pinotAppConfigs = new PinotAppConfigs(pinotConfiguration);
-    return pinotAppConfigs.toJSONString();
-  }
+  @GET
+  @Path("/appconfigs")
+  @Produces(MediaType.APPLICATION_JSON)
+  String getAppConfigs();
 }
